# public/api-docs/models/openapi.yaml
openapi: 3.0.2
info:
  title: MiW REST api
  description: '[UPM] MiW REST api'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.3
  contact:
    name: Máster Universitario en Ingeniería Web
    url: http://miw.etsisi.upm.es/
    email: miw.etsisi@upm.es
servers:
  - url: '{schema}://{servername}:{port}/api/v1'
    description: 'MiW Production server (uses live data)'
    variables:
      schema:
        enum:
          - http
          - https
        default: https
      servername:
        default: 127.0.0.1
      port:
        default: '8000'
...
paths:
...
  /results:
    get:
      tags: [ Results ]
      summary: Retrieves the collection of Result resources.
      description: 'Returns all results from the system that the user has access to.'
      operationId: api_results_cget
      security:
        - MiWApiSecurity: []
      responses:
        '200':
          description: 'Array of results'
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: { description: 'Results array', type: array, items: { $ref: '#/components/schemas/Result' } }
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
    post:
      tags: [ Results ]
      summary: Creates a Result resource.
      description: 'Creates a new result'
      operationId: api_results_post
      security:
        - MiWApiSecurity: []
      requestBody:
        $ref: '#/components/requestBodies/ResultRequestBody'
      responses:
        '201':
          description: '`Created`: result created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          headers:
            Location:
              $ref: '#/components/headers/Location'
        '400':
          $ref: '#/components/responses/400_BadRequest_Response'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '422':
          $ref: '#/components/responses/422_UnprocessableEntity_Response'
    options:
      tags: [ Results ]
      summary: Provides the list of HTTP supported methods.
      description: 'Return a `Allow` header with a comma separated list of HTTP supported methods.'
      operationId: api_results_coptions
      responses:
        '204':
          description: '`Allow` header &lt;Response body is empty&gt;'
          headers:
            Allow:
              $ref: '#/components/headers/Allow'
  '/results/{userId}/users':
    put:
      tags: [ Results ]
      summary: Clone results into another user
      description: 'Clone results into another user'
      operationId: api_clone_results
      security:
        - MiWApiSecurity: [ ]
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        $ref: '#/components/requestBodies/CloneRequestBody'
      responses:
        '200':
          description: 'Results cloned'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '400':
          $ref: '#/components/responses/400_BadRequest_Response'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
    get:
      tags: [ Results ]
      summary: Retrieves all results for one user
      description: 'Retrieves all results for one user'
      operationId: api_results_get_by_user
      security:
        - MiWApiSecurity: [ ]
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
  '/results/{resultId}':
    get:
      tags: [ Results ]
      summary: Retrieves a Result resource based on a single ID.
      description: 'Returns the result identified by `resultId`.'
      operationId: api_results_get
      security:
        - MiWApiSecurity: [ ]
      parameters:
        - $ref: '#/components/parameters/resultId'
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
    put:
      tags:
        - Results
      summary: Updates the Result resource.
      description: 'Updates the result identified by `resultId`.'
      operationId: api_results_put
      security:
        - MiWApiSecurity: [ ]
      parameters:
        - $ref: '#/components/parameters/resultId'
      requestBody:
        $ref: '#/components/requestBodies/ResultRequestBody'
      responses:
        '209':
          description: '`Content Returned`: result previously existed and is now updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '400':
          $ref: '#/components/responses/400_BadRequest_Response'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
    delete:
      tags: [ Results ]
      summary: Removes the Result resource.
      description: 'Deletes the result identified by `resultId`.'
      operationId: api_results_delete
      security:
        - MiWApiSecurity: [ ]
      parameters:
        - $ref: '#/components/parameters/resultId'
      responses:
        '204':
          description: 'User deleted &lt;Response body is empty&gt;'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
    options:
      tags: [ Results ]
      summary: Provides the list of HTTP supported methods.
      description: 'Return a `Allow` header with a comma separated list of HTTP supported methods.'
      operationId: api_results_options_id
      parameters:
        - $ref: '#/components/parameters/resultId'
      responses:
        '204':
          description: '`Allow` header &lt;Response body is empty&gt;'
          headers:
            Allow:
              $ref: '#/components/headers/Allow'
components:
  headers:
    ETag:
      description: Entity Tag (unique identifier provided for cache validation)
      schema: { type: string }
    Location:
      description: "`URI` of the new resource"
      schema: { type: string, format: url }
    Allow:
      description: List of HTTP supported methods
      schema: { type: string }
  schemas:
    Result:
      type: object
      properties:
        id:
          description: 'Result ID'
          type: integer
          format: int64
        result:
          description: 'Result Points'
          type: integer
          format: int64
        user_id:
          description: 'User that belongs the result'
          type: integer
          format: int64
        time:
          description: 'Result Time'
          type: string
          format: datetime
        _links:
          description: link relations
          type: object
          properties:
            parent:
              type: object
              properties: {"href": { "type": "string", "format": "url" }}
            self:
              type: object
              properties: {"href": { "type": "string", "format": "url" }}
      example:
        result:
          id: 1
          result: 100
          user_id: 1508
          time: "2020-12-08T11:48:44+00:00"
          _links:
            parent:
              href: /api/v1/results
            self:
              href: /api/v1/results/1
...
    Message:
      type: object
      description: HTTP response message
      properties:
        code:
          description: 'Response code'
          type: integer
          format: int64
        message:
          description: 'Response message'
          type: string
      example:
        code: 200
        message: Ok
  responses:
    400_BadRequest_Response:
      description: "`BAD REQUEST`: e-mail already exists"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-400:
              $ref: '#/components/examples/response-400'
    401_Unauthorized_Response:
      description: '`Unauthorized`: invalid Authorization header'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-401:
              $ref: '#/components/examples/response-401'
    403_Forbidden_Response:
      description: '`Forbidden`: you don''t have permission to access'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-403:
              $ref: '#/components/examples/response-403'
    404_Resource_Not_Found_Response:
      description: '`Not found`: resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-404:
              $ref: '#/components/examples/response-404'
    422_UnprocessableEntity_Response:
      description: "`UNPROCESSABLE ENTITY`: e-mail or password is left out"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-422:
              $ref: '#/components/examples/response-422'
  requestBodies:
...
    CloneRequestBody:
      description: 'data for Clone'
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              parent:
                description: 'user from which results are cloned'
                type: integer
                format: int64
    ResultRequestBody:
      description: '`Result` data'
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              result:
                description: 'Result points'
                type: integer
                format: int64
              time:
                description: 'Result time'
                type: string
                format: date-time
...
  parameters:
    userId:
      name: userId
      in: path
      description: 'ID of user'
      required: true
      schema: {"type": "integer", pattern: '^\d+$'}
    resultId:
      name: resultId
      in: path
      description: 'ID of result'
      required: true
      schema: {"type": "integer", pattern: '^\d+$'}
  examples:
    response-400:
      value:
        code: 400
        message: 'BAD REQUEST: e-mail already exists'
    response-401:
      value:
        code: 401
        message: 'UNAUTHORIZED: invalid Authorization header'
    response-403:
      value:
        code: 403
        message: "FORBIDDEN You don't have permission to access"
    response-404:
      value:
        code: 404
        message: 'NOT FOUND: Resource not found'
    response-422:
      value:
        code: 422
        message: 'UNPROCESSABLE ENTITY: e-mail or password is left out'
  securitySchemes:
    MiWApiSecurity:
      type: http
      bearerFormat: JWT
      scheme: bearer